git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@476 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/source/MeshAlgAdjacency.cpp
@@ -620,7 +620,7 @@ void MeshAlg::weldAdjacency(
     Array<Face>&          faceArray,
     Array<Edge>&          edgeArray,
     Array<Vertex>&        vertexArray,
-    double                radius) {
+    float                 radius) {
 
     // Num vertices
     const int n = originalGeometry.size();
